Rename containers in pages folder and rename page index.js files
