+++ src/containers/OperatorResizableSectionContainer/OperatorResizableSectionContainer.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 
 import { PropertiesPanel } from 'components';
-import DatasetDrawerContainer from 'pages/Experiments/Experiment/Drawer/DatasetDrawer/Container';
-import GenericDrawerContainer from 'pages/Experiments/Experiment/Drawer/GenericDrawer/Container';
+import DatasetDrawerContainer from 'pages/Experiments/Experiment/Drawer/DatasetDrawer/DatasetDrawerContainer';
+import GenericDrawerContainer from 'pages/Experiments/Experiment/Drawer/GenericDrawer/GenericDrawerContainer';
 
 import './OperatorResizableSectionContainer.less';
