+++ src/pages/App/App.page.jsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import { Provider } from 'react-redux';
+import PT_BR from 'antd/es/locale/pt_BR';
+import { Layout, ConfigProvider } from 'antd';
+
+import Store from 'store';
+import { MainRoutes } from 'routes';
+import MainSider from 'components/MainSider';
+import { CustomConfirmRouterContainer } from 'containers';
+
+import './App.style.less';
+
+const App = () => (
+  <Provider store={Store}>
+    <CustomConfirmRouterContainer>
+      <ConfigProvider locale={PT_BR}>
+        <Layout className='main-layout'>
+          <MainSider />
+
+          <Layout>
+            <MainRoutes />
+          </Layout>
+        </Layout>
+      </ConfigProvider>
+    </CustomConfirmRouterContainer>
+  </Provider>
+);
+
+export default App;
