--- src/pages/Deployment/Deployment.md
@@ -1,12 +1,0 @@
-<!-- **Padrão:**
-
-Renderização padrão da página.
-```js
-// 480p resolution
-const screeWidth = '720px';
-const screenHeight = '480px';
-
-<div style={{ width: screeWidth, height: screenHeight }}>
-  <Deployment/>
-</div>
-``` -->
