+++ src/pages/JupyterLab/JupyterLab.page.jsx
@@ -2,11 +2,8 @@ import React from 'react';
 
 import JupyterLabContainer from 'containers/JupyterLabContainer';
 
-/**
- * Pagina de pré-loading do JupyterLab.
- */
-function JupyterLab() {
+const JupyterLab = () => {
   return <JupyterLabContainer />;
-}
+};
 
 export default JupyterLab;
