+++ src/pages/ProjectDetails/index.js
@@ -1,30 +1,3 @@
-// CORE LIBS
-import React from 'react';
-import { Layout } from 'antd';
+import ProjectDetails from './ProjectDetails.page';
 
-// CONTAINERS
-import HeaderProjectDetailsContainer from 'containers/HeaderProjectDetailsContainer/index';
-import TasksMenuDetailsContainer from 'containers/TasksMenuDetailsContainer/index';
-import ProjectDetailsContainer from 'containers/ProjectDetailsContainer/index';
-
-import './style.less';
-
-const { Sider, Content } = Layout;
-
-const ProjectsDetailsContent = () => (
-  <>
-    <HeaderProjectDetailsContainer />
-    <Layout style={{ overflow: 'hidden' }}>
-      <Sider width='20%'>
-        <TasksMenuDetailsContainer />
-      </Sider>
-      <Content>
-        <Layout style={{ height: '100%' }} className='content'>
-          <ProjectDetailsContainer />
-        </Layout>
-      </Content>
-    </Layout>
-  </>
-);
-
-export default ProjectsDetailsContent;
+export default ProjectDetails;
