+++ src/pages/Projects/ProjectsTable/ProjectsTableContainer.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 
 // COMPONENTS
-import ProjectsTablePagination from '../ProjectsTablePagination/Container';
+import ProjectsTablePagination from '../ProjectsTablePagination/ProjectsTablePaginationContainer';
 import ProjectsTable from './index';
 import { MyProjectsEmptyPlaceholder } from 'components/EmptyPlaceholders';
