+++ src/pages/index.js
@@ -1,17 +1,17 @@
-import Deployment from './Deployment';
-import JupyterLab from './JupyterLab';
-import Projects from './Projects';
-import ProjectsDetails from './ProjectDetails';
-import Experiments from './Experiments';
 import Tasks from './Tasks';
 import Error404 from './Error404';
+import Projects from './Projects';
+import Deployment from './Deployment';
+import JupyterLab from './JupyterLab';
+import Experiments from './Experiments';
+import ProjectsDetails from './ProjectDetails';
 
 export {
-  Deployment,
-  JupyterLab,
-  Projects,
-  ProjectsDetails,
-  Experiments,
   Tasks,
   Error404,
+  Projects,
+  Deployment,
+  JupyterLab,
+  Experiments,
+  ProjectsDetails,
 };
